% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cols2DocTbl.R
\name{cols2DocTbl}
\alias{cols2DocTbl}
\title{Build Empty 'Document' Table}
\usage{
cols2DocTbl(cols)
}
\arguments{
\item{cols}{named character vector}
}
\value{
data.frame
}
\description{
Initialize an empty data.frame of character vectors. Object is meant
as a base class implementation of a requirements document table.
}
