% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-docTemplates.R
\name{data2DocTemplate}
\alias{data2DocTemplate}
\title{Convert Tablular Data to Document Template Format}
\usage{
data2DocTemplate(data, template)
}
\arguments{
\item{data}{data.frame}

\item{template}{character}
}
\value{
R Object
}
\description{
Convert Tablular Data to Document Template Format
}
