% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class_docFuncReq.R
\name{docFuncReq}
\alias{docFuncReq}
\title{Constructor with Validation for S3 Class 'docFuncReq'}
\usage{
docFuncReq(bool_out = FALSE, throw_err = TRUE)
}
\arguments{
\item{bool_out}{TRUE/FALSE - should output be TRUE when input object
is valid, or should the input object be silently returned.}

\item{throw_err}{TRUE/FALSE - should function stop and throw error
messages in case of invalid input objects, or return character vector
of messages as function output.}
}
\value{
S3 object of class 'docFuncReq'
}
\description{
Constructor with Validation for S3 Class 'docFuncReq'
}
