% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docTemplates}
\alias{docTemplates}
\title{Requirement Document Templates}
\format{
A data frame with 21 rows and 3 columns:
\describe{
  \item{template}{(character) name of document}
  \item{name}{(character) name of field within document}
  \item{colstring}{(character) user-friendly colstring}
  ...
}
}
\usage{
docTemplates
}
\description{
A data.frame of definitions for defined document templates
}
\keyword{datasets}
