% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generic_exportExcel.R
\name{exportExcel.docTestCases}
\alias{exportExcel.docTestCases}
\title{S3 Method - Export S3 Class 'docTestCases' to Excel}
\usage{
\method{exportExcel}{docTestCases}(obj, file, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{file}{character - file path for excel export}

\item{...}{R ellipsis}
}
\description{
S3 Method - Export S3 Class 'docTestCases' to Excel
}
