% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generic_exportExcel.R
\name{exportExcel.docUserStory}
\alias{exportExcel.docUserStory}
\title{S3 Method - Export S3 Class 'docUserStory' to Excel}
\usage{
\method{exportExcel}{docUserStory}(obj, file, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{file}{character - file path for excel export}

\item{...}{R ellipsis}
}
\description{
S3 Method - Export S3 Class 'docUserStory' to Excel
}
