% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-docTemplates.R
\name{getDocTemplate}
\alias{getDocTemplate}
\title{Get Document Template Fields}
\usage{
getDocTemplate(template, cols)
}
\arguments{
\item{template}{character - name of template}

\item{cols}{- character - column names to return}
}
\value{
R Object
}
\description{
Get Document Template Fields
}
