% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_getDocTemplates.R
\name{getDocTemplates}
\alias{getDocTemplates}
\title{Get Document Template Fields}
\usage{
getDocTemplates(template, cols)
}
\arguments{
\item{template}{character - name of template}

\item{cols}{- character - column names to return}
}
\value{
R Object
}
\description{
Get Document Template Fields
}
