% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_action_tbl.R
\name{mod_action_tbl_server}
\alias{mod_action_tbl_server}
\title{Module Server - Action Table}
\usage{
mod_action_tbl_server(id, docs, doc_id, rownames = FALSE)
}
\arguments{
\item{id}{character - module id}

\item{docs}{reactive values - list of reactive values created with 'shiny::reactiveValues()'}

\item{doc_id}{character - name of element in 'docs' containing table data}

\item{rownames}{logical - TRUE/FALSE, specifies if rownames should be displayed in table}
}
\description{
Server Function for R shiny module 'action_tbl'.
Uses 'bslib' and 'DT' packages to render an editable table.
}
