% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_action_tbl.R
\name{mod_action_tbl_ui}
\alias{mod_action_tbl_ui}
\title{Module UI - Action Table}
\usage{
mod_action_tbl_ui(id, title, position = "left")
}
\arguments{
\item{id}{character - module id}

\item{title}{character - custom title for table}

\item{position}{character - 'left'/'right', specifies position of sidebar}
}
\description{
UI Function for R shiny module 'action_tbl'.
Uses 'bslib' and 'DT' packages to render an editable table.
}
