% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-zzz_helpers.R
\name{msg_setdiff}
\alias{msg_setdiff}
\title{User-Friendly Error Message for Set Differences}
\usage{
msg_setdiff(x, y)
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}
}
\value{
character
}
\description{
Displays elements missing from 'x' and unexpected elements in 'y'
}
