% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-docFuncReq.R
\name{new_docFuncReq}
\alias{new_docFuncReq}
\title{Constructor for S3 Class 'docFuncReq'}
\usage{
new_docFuncReq()
}
\value{
S3 object of class 'docFuncReq'
}
\description{
Construct a new instance of an S3 object 'docFuncReq'.
Inherits functionality from S3 Class 'docTbl'.
}
