% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-docTestCases.R
\name{new_docTestCases}
\alias{new_docTestCases}
\title{Constructor for S3 Class 'docTestCases'}
\usage{
new_docTestCases()
}
\value{
S3 object of class 'docTestCases'
}
\description{
Construct a new instance of an S3 object 'docTestCases'.
Inherits functionality from S3 Class 'docTbl'.
}
