% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class_docUserStory.R
\name{new_docUserStory}
\alias{new_docUserStory}
\title{Constructor for S3 Class 'docUserStory'}
\usage{
new_docUserStory()
}
\value{
S3 object of class 'docUserStory'
}
\description{
Construct a new instance of an S3 object 'docUserStory'.
Inherits functionality from S3 Class 'docTbl'.
}
