% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-zzz_helpers.R
\name{obj_msg_result}
\alias{obj_msg_result}
\title{Product Final Output based on User Settings and Error Messages}
\usage{
obj_msg_result(obj, msg, bool_out = FALSE, throw_err = TRUE)
}
\arguments{
\item{obj}{R Object}

\item{msg}{character - error messages}

\item{bool_out}{TRUE/FALSE - should output be TRUE when input object
is valid, or should the input object be silently returned.}

\item{throw_err}{TRUE/FALSE - should function stop and throw error
messages in case of invalid input objects, or return character vector
of messages as function output.}
}
\value{
R Object (see parameters)
}
\description{
Product Final Output based on User Settings and Error Messages
}
