% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-classes.R
\name{validate_docUserStory}
\alias{validate_docUserStory}
\title{Validator for S3 Class 'docUserStory'}
\usage{
validate_docUserStory(obj, bool_out = FALSE, throw_err = TRUE)
}
\arguments{
\item{obj}{R Object}

\item{bool_out}{TRUE/FALSE - should output be TRUE when input object
is valid, or should the input object be silently returned.}

\item{throw_err}{TRUE/FALSE - should function stop and throw error
messages in case of invalid input objects, or return character vector
of messages as function output.}
}
\value{
R Object (see parameters)
}
\description{
Validator for S3 Class 'docUserStory'
}
