% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-classes.R
\name{new_docNonFuncReq}
\alias{new_docNonFuncReq}
\title{Constructor for S3 Class 'docNonFuncReq'}
\usage{
new_docNonFuncReq(data = NULL)
}
\arguments{
\item{data}{data.frame}
}
\value{
S3 object of class 'docNonFuncReq'
}
\description{
Construct a new instance of an S3 object 'docNonFuncReq'.
Inherits functionality from S3 Class 'docTbl'.
}
