% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-classes.R
\name{new_docTbl}
\alias{new_docTbl}
\title{Constructor for S3 Class 'docTbl'}
\usage{
new_docTbl(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
S3 object of class 'docTbl'
}
\description{
Construct a new instance of an S3 object 'docTbl'. Uses a character vector
to initialize a data.frame of character columns. Column names are taken
from 'names(cols)'. New attribute 'colstring' is set equal to 'cols'.
}
